'Alpha' skeleton of interface centralization have been created.
Experimental branch, which aim is to ease building of implementation.
It was decided to create a single access point to the entire template functionality with the help of a certain class type structure.
Very big changes are required, because the real implementation is extremely complicated and could lead to the formation of circular dependencies.