Interaction logic with the 'Entry' frame was described.
One more nesting level 'src/Graphics/Frames/' have been created.

And added a piece of not so pretty yet working code. (Will be refactored somewhen in the future.)
This code describes the interaction logic with the entry frame.

Pattern synonyms were added to ease country names localization.

It's time to delte the branch.
UI logic was completely refactored, but some parts of it will be improved over the time.